+++ src/util/routing.ts
@@ -201,7 +201,7 @@ class RestEntry {
         if (config.exists(REST)) {
             const rest = config.get(REST);
             if (Array.isArray(rest)) {
-                this.loadRest(config, rest.length);
+                this.loadRest(config);
                 const exact = Array.from(this.exactRoutes.keys());
                 if (exact.length > 0) {
                     log.info({'type': 'url', 'match': 'exact', 'total': exact.length, 'path': exact.sort()});
@@ -357,7 +357,10 @@ class RestEntry {
         }
     }
 
-    loadRest(config: ConfigReader, total: number): void {
+    loadRest(config: ConfigReader): void {
+        this.addDefaultEndpoints(config);
+        const rest = config.get(REST);
+        const total = rest.length;
         // perform custom sort in ascending order of URL, methods and services
         const keys = [];
         for (let i=0; i < total; i++) {
@@ -365,7 +368,7 @@ class RestEntry {
             const methods = config.get(REST+"["+i+"]."+METHODS);
             const url = config.getProperty(REST+"["+i+"]."+URL_LABEL);
             if (url && Array.isArray(methods) && (typeof services == 'string' || Array.isArray(services))) {
-                keys.push(`${url}|${JSON.stringify(methods)}|${JSON.stringify(services)}|${i}`);
+                keys.push(`${url} | ${JSON.stringify(methods)} |${i}`);
             } else {                
                 log.error(`Skip invalid REST entry ${config.get(REST+"["+i+"]")}`);
             }
@@ -375,7 +378,7 @@ class RestEntry {
         const mm = new MultiLevelMap();
         let n = 0;
         for (const k of keys) {     
-            const idx = util.str2int(k.substring(k.lastIndexOf('|') + 1));
+            const idx = k.substring(k.lastIndexOf('|') + 1);
             mm.setElement(REST+"["+n+"]", config.get(REST+"["+idx+"]"));
             n++;
         }
@@ -387,6 +390,44 @@ class RestEntry {
             if (url && Array.isArray(methods) && (typeof services == 'string' || Array.isArray(services))) {
                 this.loadRestEntry(sortedConfig, i, !url.includes("{") && !url.includes("}") && !url.includes("*"));
             }
+        }
+    }
+
+    addDefaultEndpoints(config: ConfigReader) {
+        const defaultRest = util.loadYamlFile(util.getFolder("../resources/default-rest.yaml"));
+        const defaultRestEntries = defaultRest.getElement(REST) as Array<object>;
+        const defaultTotal = defaultRestEntries.length;
+        const essentials = {};
+        const configured = new Array<string>();
+        for (let i=0; i < defaultTotal; i++) {
+            const methods = defaultRest.getElement(REST+"["+i+"]."+METHODS);
+            const url = defaultRest.getElement(REST+"["+i+"]."+URL_LABEL);
+            essentials[url+" "+JSON.stringify(methods)] = i;
+        }
+        const restEntries = config.get(REST) as Array<object>;
+        let total = restEntries.length;
+        for (let i=0; i < total; i++) {
+            const methods = config.get(REST+"["+i+"]."+METHODS);
+            const url = config.getProperty(REST+"["+i+"]."+URL_LABEL);
+            if (url && Array.isArray(methods)) {
+                configured.push(url+" "+JSON.stringify(methods));
+            }
+        }
+        // find out if there are missing default entries in the configured list
+        const missing = new Array<string>();
+        for (const entry of Object.keys(essentials)) {
+            if (!configured.includes(entry)) {
+                missing.push(entry);
+            }
+        }
+        if (missing.length > 0) {
+            const map = new MultiLevelMap(config.getMap());
+            for (const entry of missing) {
+                const idx = essentials[entry] as number;
+                map.setElement(REST + "[" + total + "]", defaultRest.getElement(REST + "[" + idx + "]"));
+                total++;
+            }
+            config.reload(map);
         }
     }